requirements.txt
Django==3.2.4
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
pytest==5.4.1
pytest-django==3.9.0
psycopg2==2.9.1
